requirement.txt
python-binance
numpy==1.26.0
TA-Lib
websocket-client